requirements.txt
numpy==2.2.6
opencv-contrib-python==4.12.0.88
ptwt==1.0.1
pytorch-msssim==1.0.0
PyYAML==6.0.3
scikit-image==0.25.2
torch==2.9.0+cu130
torchmetrics==1.8.2
torchvision==0.24.0+cu130
tqdm==4.67.1